Add support for multiple images
Add a new function that accepts multiple jpeg paths which are loaded on
demand instead of during filter creation, also add optional parameters
for setting the framerate.

Note that there is currently no error checking or attempts to conform
mismatching images in your sequence, so ensure that you only pass valid
paths to images with the same resolution.